+++ Engine/Source/Runtime/SynthBenchmark/Private/SynthBenchmarkPrivate.cpp
@@ -263,7 +263,7 @@ void FSynthBenchmark::GetRHIInfo(FGPUAdpater& Out, FString& RHIName) const
 	{
 		FTextureMemoryStats TextureMemStats;
 		RHIGetTextureMemoryStats(TextureMemStats);
-		WriteFStringToResults(Out.AdapterDedicatedMemoryMB, FString::FromInt(TextureMemStats.DedicatedVideoMemory / (1024*1024)));
+		WriteFStringToResults(Out.AdapterDedicatedMemoryMB, FString::FromInt(static_cast<int32>(TextureMemStats.DedicatedVideoMemory / (1024*1024))));
 
 		RHIName = GDynamicRHI->GetName();
 	}
